# Use const generics for downscaling