--- a/Cargo.toml
+++ b/Cargo.toml
@@ -100,6 +100,7 @@ rust_hawktracer = "0.7.0"
 arrayref = "0.3.6"
 const_fn_assert = "0.1.2"
 nom = { version = "7.0.0", optional = true }
+new_debug_unreachable = "1.0.4"
 
 [dependencies.image]
 version = "0.23"