The main bottleneck of the fast scene detection mode is the internal downscaling which is performed. This PR changes the downscaling code to be specialized for each chosen scale factor (rather than one function which handles all scale factors), which in turn enables several optimizations in the compiler that are not possible for a general downscaling function. Overall, this PR's new downscaling is over 6x faster when compiling with AVX2 enabled, benchmarked on a 1280x720 yuv420p video.

Other changes:
- Remove rayon threading from `downscale_in_place`, as it seemed to prevent auto-vectorization by the compiler
- Allocate `frame_ref_buffer` and `downscaled_frame_buffer` on the stack instead of the heap
- Remove initialization checks with `unreachable_unchecked` (which originally existed because of borrow checker limitations)

Comparison of performance:

Before this PR:

![image](https://user-images.githubusercontent.com/48274562/166078608-acc0e1ca-69ec-4687-83d4-cd5796959748.png)

After this PR:

![image](https://user-images.githubusercontent.com/48274562/166078620-5fce5d78-cb94-4ce1-a16f-65d78a9abcaf.png)

With this PR, `fast_scenecut` in total (on the tested 720p clip) is about 5.9x faster (0.798 ms vs 0.135 ms average wall duration, which amounts to a difference of 1253 fps and 7407 fps throughput).